Enable aggressive static analysis flags - resolve resulting issues
Enable -Wall -Wextra and -Werror (as well as a few more useful lints)
and then fix everything to get it compiling.

* The snprintf statements in mm_carrier.c and mm_carrier_mtrl.c seem
  to leave a space for the null terminator, but snprintf guarantees a
  null terminator, so you will get truncation!
* A number of places had for loops with `int i` but were compared
  against the output of `sizeof` - switched them to `size_t i`
* There were a number of places which used `%llx`, `%lld`, and `%llu`
  to print out uint64_t or int64_t types - switch these to their
  standard `PRIx64`, `PRId64`, and `PRIu64` counterparts
* Add in a few missing function prototypes
* Remove const from some functions which claimed their return type was
  const, which didn't make sense since the functions returned basic
  types by copy

+++ mm_serial.c
@@ -77,7 +77,7 @@ ssize_t read_serial(mm_serial_context_t *pserial_context, void *buf, size_t coun
         }
     }
     else {
-        for (int i = 0; i < count; i++) {
+        for (size_t i = 0; i < count; i++) {
             if (feof(pserial_context->bytestream)) {
                 printf("%s: Terminating due to EOF.\n", __FUNCTION__);
                 fflush(stdout);
@@ -102,7 +102,7 @@ ssize_t read_serial(mm_serial_context_t *pserial_context, void *buf, size_t coun
     }
 
     if (pserial_context->logstream != NULL) {
-        for (int i = 0; i < count; i++) {
+        for (size_t i = 0; i < count; i++) {
             fprintf(pserial_context->logstream, "UART: RX: %02X\n", ((uint8_t*)buf)[i]);
         }
     }
@@ -113,7 +113,7 @@ ssize_t write_serial(mm_serial_context_t *pserial_context, const void *buf, size
     ssize_t bytes_written = count;
 
     if (pserial_context->logstream != NULL) {
-        for (int i = 0; i < count; i++) {
+        for (size_t i = 0; i < count; i++) {
             fprintf(pserial_context->logstream, "UART: TX: %02X\n", ((uint8_t*)buf)[i]);
         }
     }
